17.03.22

+++ app/src/main/java/com/example/appen/Uv.kt
@@ -83,7 +83,7 @@ data class DetailsXXX(
     val precipitation_amount: Double,
     val precipitation_amount_max: Double,
     val precipitation_amount_min: Double,
-    val probability_of_precipitation: Int,
+    val probability_of_precipitation: Double,
     val probability_of_thunder: Double,
     val ultraviolet_index_clear_sky_max: Int
 )
@@ -94,7 +94,7 @@ data class DetailsXX(
     val precipitation_amount: Double,
     val precipitation_amount_max: Double,
     val precipitation_amount_min: Double,
-    val probability_of_precipitation: Int,
+    val probability_of_precipitation: Double,
     val probability_of_thunder: Double,
     val ultraviolet_index_clear_sky_max: Int
 )
@@ -105,7 +105,7 @@ data class DetailsX(
     val precipitation_amount: Double,
     val precipitation_amount_max: Double,
     val precipitation_amount_min: Double,
-    val probability_of_precipitation: Int,
+    val probability_of_precipitation: Double,
     val probability_of_thunder: Double,
     val ultraviolet_index_clear_sky_max: Int
 )
